requirements.txt
colorama
detectree
fastapi[standard]
ipykernel
labelme
matplotlib
numpy
Pillow
psycopg2
pytest
requests
setuptools
sqlalchemy
torch
torchvision
tqdm